# BackendServices/matrimonialwebapi/src/MatrimonialApi/wwwroot/swagger-original.json
{
  "openapi" : "3.0.3",
  "info" : {
    "title" : "Matrimonial API - OpenAPI 3.0",
    "description" : "Design and definition of Matrimonial APIs created for practice and teaching",
    "contact" : {
      "email" : "floatingrays@gmail.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version" : "1.0.11"
  },
...
  "servers" : [ {
    "url" : "http://localhost:3002/api"
  } ],
...
  "paths" : {
    "/admin" : {
      "summary" : "Add a admin to the system",
      "description" : "Add a admin to the system",
      "get" : {
        "tags" : [ "Admin" ],
        "summary" : "Get all admins",
        "description" : "Get all the admins present in the system",
        "parameters" : [ {
          "name" : "X-Request-Auth",
          "in" : "header",
          "description" : "Authorization token value",
          "required" : false,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/CreatedAdmin"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid input"
          },
          "422" : {
            "description" : "Validation exception"
          },
          "500" : {
            "description" : "An error occured while processing the request."
          },
          "default" : {
            "description" : "Default error"
          }
        },
        "deprecated" : false,
        "security" : [ {
          "BearerAuth" : [ ]
        } ]
      },
      "post" : {
        "tags" : [ "Admin" ],
        "summary" : "Add a admin to the system",
        "description" : "Add a admin to the system",
        "operationId" : "AddAdmin",
        "parameters" : [ {
          "name" : "X-Request-Auth",
          "in" : "header",
          "description" : "Authorization token value",
          "required" : false,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "description" : "Create a new admin",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Admin"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CreatedAdmin"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid input"
          },
          "422" : {
            "description" : "Validation exception"
          },
          "500" : {
            "description" : "An error occured while processing the request."
          },
          "default" : {
            "description" : "Default error"
          }
        },
        "deprecated" : false,
        "security" : [ {
          "BearerAuth" : [ ]
        } ]
      }
    },
    "/admin/{adminId}" : {
      "summary" : "Manage a admin from the system",
      "description" : "Manage a admin from the system",
      "get" : {
        "tags" : [ "Admin" ],
        "summary" : "Get a admin from the system",
        "description" : "Get a admin to the system",
        "operationId" : "GetAdmin",
        "parameters" : [ {
          "name" : "adminId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "X-Request-Auth",
          "in" : "header",
          "description" : "Authorization token value",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CreatedAdmin"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid input"
          },
          "422" : {
            "description" : "Validation exception"
          },
          "500" : {
            "description" : "An error occured while processing the request."
          },
          "default" : {
            "description" : "Default error"
          }
        },
        "deprecated" : false,
        "security" : [ {
          "BearerAuth" : [ ]
        } ]
      },
      "put" : {
        "tags" : [ "Admin" ],
        "summary" : "Update a admin",
        "description" : "Update a admin",
        "operationId" : "UpdateAdmin",
        "parameters" : [ {
          "name" : "adminId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "X-Request-Auth",
          "in" : "header",
          "description" : "Authorization token value",
          "required" : false,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "description" : "Update a admin",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Admin"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Successful operation"
          },
          "400" : {
            "description" : "Invalid input"
          },
          "422" : {
            "description" : "Validation exception"
          },
          "500" : {
            "description" : "An error occured while processing the request."
          },
          "default" : {
            "description" : "Default error"
          }
        },
        "deprecated" : false
      },
      "delete" : {
        "tags" : [ "Admin" ],
        "summary" : "Delete a admin",
        "description" : "Delete a admin",
        "operationId" : "DeleteAdmin",
        "parameters" : [ {
          "name" : "adminId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "X-Request-Auth",
          "in" : "header",
          "description" : "Authorization token value",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation"
          },
          "400" : {
            "description" : "Invalid input"
          },
          "422" : {
            "description" : "Validation exception"
          },
          "500" : {
            "description" : "An error occured while processing the request."
          },
          "default" : {
            "description" : "Default error"
          }
        },
        "deprecated" : false
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "Admin" : {
        "required" : [ "emailId", "firstName", "lastName" ],
        "type" : "object",
        "properties" : {
          "firstName" : {
            "type" : "string",
            "format" : "Alphabets",
            "example" : "Ram"
          },
          "middleName" : {
            "type" : "string",
            "format" : "Alphabets",
            "example" : "kumar"
          },
          "lastName" : {
            "type" : "string",
            "format" : "Alphabets",
            "example" : "Sanatani"
          },
          "emailId" : {
            "type" : "string",
            "format" : "Alpha-Numeric-Special Character@domain.name",
            "example" : "Ram@gmail.com"
          }
        },
        "xml" : {
          "name" : "admin"
        }
      },
      "CreatedAdmin" : {
        "type" : "object",
        "properties" : {
          "firstName" : {
            "type" : "string",
            "format" : "Alphabets",
            "example" : "Ram"
          },
          "middleName" : {
            "type" : "string",
            "format" : "Alphabets",
            "example" : "kumar"
          },
          "lastName" : {
            "type" : "string",
            "format" : "Alphabets",
            "example" : "Sanatani"
          },
          "emailId" : {
            "type" : "string",
            "format" : "Alpha-Numeric-Special Character@domain.name",
            "example" : "Ram@gmail.com"
          },
          "isActive" : {
            "type" : "boolean",
            "format" : "true or false",
            "example" : true
          },
          "isSuperAdmin" : {
            "type" : "boolean",
            "format" : "true or false",
            "example" : true
          },
          "createdDateTime" : {
            "type" : "string",
            "format" : "date-time"
          },
          "accountLocked" : {
            "type" : "boolean",
            "format" : "true or false",
            "example" : false
          },
          "passwordResetDateTime" : {
            "type" : "string",
            "format" : "date-time"
          }
        },
        "example" : {
          "accountLocked" : false,
          "firstName" : "Ram",
          "lastName" : "Sanatani",
          "passwordResetDateTime" : "2000-01-23T04:56:07.000+00:00",
          "createdDateTime" : "2000-01-23T04:56:07.000+00:00",
          "middleName" : "kumar",
          "emailId" : "Ram@gmail.com",
          "isSuperAdmin" : true,
          "isActive" : true
        },
        "xml" : {
          "name" : "CreatedAdmin"
        }
      },
...
    },
    "securitySchemes" : {
      "BasicAuth" : {
        "type" : "http",
        "scheme" : "basic"
      },
      "BearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      },
      "SSO_auth" : {
        "type" : "oauth2",
        "flows" : {
          "implicit" : {
            "authorizationUrl" : "https://localhost:3002/oauth/authorize",
            "scopes" : {
              "write:pets" : "modify data",
              "read:pets" : "read data"
            }
          }
        }
      },
      "api_key" : {
        "type" : "apiKey",
        "name" : "api_key",
        "in" : "header"
      }
    }
  }
}